The project is running on UTC while our time zone is IST (Indian Standard Time ) thats why its giving problem bcz the server we are using is GCP in mumbai.
I also want that the meals will automatically freeze according to the confirmation_deadline_hours in the meal_timmings table

In case of parent booking for the student .After the meal and the number of parents in case of students is slected .Instead of payment gateway a new qr will open and when the user_type -> mess_staff scans that unique qr which is generated only for that time then in mess_staff screen it will ask that confirm payment recieved or not recieved .if the mess_staff confirm the payment_recieved the users._id for that mess_staff and the amount_collected for that meal is stored to keep track of who collected theb amount for that booking and the student who booked that meal confimration will be done for the parents.

There must be a table hostel_master for students every user has user_type incase of user_type student there is table student_details which contains object hostel_details inisde that object we have hostel_name field .The hostel_master have hostel_name and mess_name .The student of a particular hostel will attend the meal only in the mess tagged with that that hostel.

The user_type mess_staff is also tagged with mess_name only the mess_staff of the particular mess can scan the qr code for the student of that mess or book the parent booking for the student of that mess.Otherwise the error recived as student is not of your mess .

The mess_staff of a particular mess will recived the meals attended , meals confirmed for that mess on its admin dashboard notr of any other mess.